Fix circular dependency warning from rollup
I noticed that when rollup builds the bundles it was emitting a circular
dependency warning for Epic -> EpicWithSpecialHeader -> Epic.
EpicWithSpecialHeader is now entirely defined by the Epic component so
it makes sense to move the HeaderSection into the Epic folder - that's
where it's used. This resolved the circular dependency.
